+++ R/lm_betaselect_methods.R
@@ -798,7 +798,7 @@ anova.glm_betaselect <- function(object,
 #'
 #' @param ci Logical. Whether
 #' confidence intervals are computed.
-#' Default is `FALSE.`
+#' Default is `TRUE`.
 #'
 #' @param level The level of confidence,
 #' default is .95, returning the 95%
@@ -832,6 +832,21 @@ anova.glm_betaselect <- function(object,
 #' output *after* selected
 #' variables standardized are returned.
 #' Default is `"beta"`.
+#'
+#' @param print_raw Control whether
+#' the estimates before selected
+#' standardization are printed when
+#' `type` is `"beta"` or `"standardized"`.
+#' If `"none"`, the default, then it
+#' will not be printed. If set to `"before_ci"`
+#' and `ci` is `TRUE`, then will be
+#' inserted to the left of the confidence
+#' intervals. If set to "after_ci"` and `ci`
+#' is `TRUE`, then will
+#' be printed to the right of the confidence
+#' intervals. If `ci` is `FALSE`, then will
+#' be printed to the right of the
+#' standardized estimates.
 #'
 #' @param ...  Additional arguments
 #' passed to other methods.
@@ -870,7 +885,7 @@ summary.lm_betaselect <- function(object,
                                   symbolic.cor = FALSE,
                                   se_method = c("boot", "bootstrap",
                                                 "t", "lm", "ls"),
-                                  ci = FALSE,
+                                  ci = TRUE,
                                   level = .95,
                                   boot_type = c("perc", "bc"),
                                   boot_pvalue_type = c("asymmetric", "norm"),
@@ -878,9 +893,11 @@ summary.lm_betaselect <- function(object,
                                            "standardized",
                                            "raw",
                                            "unstandardized"),
+                                  print_raw = c("none", "before_ci", "after_ci"),
                                   ...) {
     se_method <- match.arg(se_method)
     type <- match.arg(type)
+    print_raw <- match.arg(print_raw)
     se_method <- switch(se_method,
                         boot = "boot",
                         bootstrap = "boot",
@@ -961,6 +978,21 @@ summary.lm_betaselect <- function(object,
         out_coef <- out$coefficients
         out_coef <- cbind(out_coef[, seq_len(i), drop = FALSE],
                           out_ci,
+                          out_coef[, -seq_len(i), drop = FALSE])
+        out$coefficients <- out_coef
+      }
+    if (print_raw != "none") {
+        b_ustd <- stats::coef(object$lm_betaselect$ustd)
+        if (ci) {
+            i <- switch(print_raw,
+                        before_ci = which(colnames(out$coefficients) == "CI.Lower") - 1,
+                        after_ci = which(colnames(out$coefficients) == "CI.Upper"))
+          } else {
+            i <- which(colnames(out$coefficients) == "Estimate")
+          }
+        out_coef <- out$coefficients
+        out_coef <- cbind(out_coef[, seq_len(i), drop = FALSE],
+                          "Raw" = b_ustd,
                           out_coef[, -seq_len(i), drop = FALSE])
         out$coefficients <- out_coef
       }
@@ -1088,6 +1120,11 @@ print.summary.lm_betaselect <- function(x,
                 beta = "- Results *after* standardization are reported.",
                 raw = "- Results *before* standardization are reported."),
                 exdent = 2))
+    if ("Raw" %in% colnames(x$coefficients)) {
+        tmp <- c(tmp,
+                strwrap("- 'Raw' shows the estimates *before* standardization.",
+                        exdent = 2))
+      }
     if (x$lm_betaselect$se_method == "boot") {
         tmp <- c(tmp,
                  strwrap("- Nonparametric bootstrapping conducted.",